example-3/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.32
Werkzeug==2.2.3
